review: rename font_idx → font_index, drop no-op cast
Two cleanups per architecture-guardian:
- Rename make_body_format's first parameter font_idx → font_index
  for alignment with the Page::Format field of the same name.
  Designated-init reads uniformly across declaration and call.
- Drop a no-op (int16_t) cast at page_cache.cpp:707 — vestigial
  from the old (int8_t) narrowing in the pre-refactor struct
  literal; misleads readers into thinking there's a type bridge.

No behavior change.

+++ src/models/page_cache.cpp
@@ -704,8 +704,8 @@ bool render_page_into(uint8_t * fb, const PageLocs::PageId & page_id)
   if (font_idx == -1) font_idx = 3;
   int8_t font_size = epub.get_book_format_params()->font_size;
 
-  Page::Format fmt = Page::make_body_format((int16_t) font_idx, font_size,
-                                             page_top, page_bottom);
+  Page::Format fmt = Page::make_body_format(font_idx, font_size,
+                                            page_top, page_bottom);
 
   DOM dom;
   // The Page singleton is shared with book_viewer's foreground
